# spec/openapi.yml
...
openapi: 3.0.2
info:
  title: Marquez
  version: 0.19.2-SNAPSHOT
  description: Marquez is an open source **metadata service** for the **collection**, **aggregation**,
    and **visualization** of a data ecosystem's metadata.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local API server
paths:
...
  /sources/{source}:
    parameters:
      - $ref: '#/components/parameters/source'
    put:
      summary: Create a source
      description: Creates a new source object. A source is the physical location of a dataset such as
        a table in PostgreSQL, or topic in Kafka. A source enables the grouping of physical datasets
        to their physical source.
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedSource'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
    get:
      summary: Retrieve a source
      description: Returns a source.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /sources:
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all sources
      description: Returns a list of sources.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
  /namespaces/{namespace}/datasets/{dataset}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
    put:
      summary: Create a dataset
      description: Creates a new dataset.
      tags:
        - Datasets
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/DbTable'
                - $ref: '#/components/schemas/Stream'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
    get:
      summary: Retrieve a dataset
      description: Returns a dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /namespaces/{namespace}/datasets/{dataset}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
      - $ref: '#/components/parameters/version'
    get:
      summary: Retrieve a version for a dataset
      description: Returns a version for a dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetVersion'
  /namespaces/{namespace}/datasets/{dataset}/versions:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
    get:
      summary: List all versions for a dataset
      description: Returns a list of versions for a dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetVersionList'
  /namespaces/{namespace}/datasets:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all datasets
      description: Returns a list of datasets.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
  /namespaces/{namespace}/datasets/{dataset}/tags/{tag}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
      - $ref: '#/components/parameters/tag'
    post:
      summary: Tag a dataset
      description: Tag an existing dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /namespaces/{namespace}/datasets/{dataset}/fields/{field}/tags/{tag}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
      - $ref: '#/components/parameters/field'
      - $ref: '#/components/parameters/tag'
    post:
      summary: Tag a field
      description: Tag an existing field of a dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /namespaces/{namespace}/jobs/{job}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    put:
      summary: Create a job
      description: Creates a new job object. All job objects are immutable and are uniquely identified by a generated ID.
        Marquez will create a version of a job each time the contents of the object is modified. For example, the `location`
        of a job may change over time resulting in new versions. The accumulated versions can be listed, used to rerun a
        specific job version or possibly help debug a failed job run.
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedJob'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
    get:
      summary: Retrieve a job
      description: Retrieve a job.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  /namespaces/{namespace}/jobs:
    parameters:
      - $ref: '#/components/parameters/namespace'
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all jobs
      description: Returns a list of jobs.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
  /namespaces/{namespace}/jobs/{job}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
      - $ref: '#/components/parameters/version'
    get:
      summary: Retrieve a version for a job
      description: Returns a version for a job.
      tags:
        - Jobs
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobVersion'
  /namespaces/{namespace}/jobs/{job}/versions:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    get:
      summary: List all versions for a job
      description: Returns a list of versions for a job.
      tags:
        - Jobs
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobVersionList'
  /namespaces/{namespace}/jobs/{job}/runs:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    post:
      summary: Create a run
      description: Creates a new run object for a job.
      deprecated: true
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedRun'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
    get:
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all runs
      description: Returns a list of runs for a job.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
  /jobs/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/runId'
    get:
      summary: Retrieve a run
      description: Retrieve a run.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncompleteRun'
...
components:
  parameters:
...
    namespace:
      name: namespace
      in: path
      description: The name of the namespace.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-namespace
    source:
      name: source
      in: path
      description: The name of the source.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-source
    dataset:
      name: dataset
      in: path
      description: The name of the dataset.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-dataset
    tag:
      name: tag
      in: path
      description: The name of the tag.
      required: true
      schema:
        type: string
        example: SENSITIVE
    field:
      name: field
      in: path
      description: The name of the field.
      required: true
      schema:
        type: string
        example: my_field
    job:
      name: job
      in: path
      description: The name of the job.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-job
...
    version:
      name: version
      in: path
      description: The ID of the job or dataset version.
      required: true
      schema:
        type: string
        format: uuid
        example: ea9badc5-7cb2-49af-9a9f-155771d3a797
    limit:
      name: limit
      in: query
      description: The number of results to return from offset
      required: false
      schema:
        type: integer
        example: 25
        default: 100
    offset:
      name: offset
      in: query
      description: The initial position from which to return results
      required: false
      schema:
        type: integer
        default: 0
...
  schemas:
...
    CreatedSource:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the source.
          type: string
      required:
        - type
        - connectionUrl
      example:
        type: POSTGRESQL
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    Source:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        name:
          description: The name of the source.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was updated.
          type: string
          format: date-time
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the dataset.
          type: string
      example:
        type: POSTGRESQL
        name: my-source
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    SourceList:
      type: object
      properties:
        sources:
          items:
            $ref: '#/components/schemas/Source'
...
    DbTable:
      type: object
      properties:
        type:
          description: The type of the dataset.
          type: string
          enum: [DB_TABLE]
        physicalName:
          description: The **physical** name of the table.
          type: string
        sourceName:
          description: The name of the source associated with the table.
          type: string
        fields:
          description: The fields of the table.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The data type of the field.
                type: string
              tags:
                description: List of tags.
                type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: array
          items:
            type: string
        description:
          description:  The description of the table.
          type: string
        runId:
          description: The ID associated with the run modifying the table.
          type: string
      required:
        - type
        - physicalName
        - sourceName
        - fields
      example:
        type: DB_TABLE
        physicalName: public.mytable
        sourceName: my-source
        fields: [
          {name: a, type: INTEGER},
          {name: b, type: TIMESTAMP},
          {name: c, type: INTEGER},
          {name: d, type: INTEGER}
        ]
        description: "My first dataset!"
    Stream:
      type: object
      properties:
        type:
          description: The type of the dataset.
          type: string
          enum: [STREAM]
        physicalName:
          description: The **physical** name of the stream.
          type: string
        sourceName:
          description: The name of the source associated with the stream.
          type: string
        schemaLocation:
          description: The URL of the schema for the stream data.
          type: string
        fields:
          description: The schema fields for the stream data.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The type of the field.
                type: string
              tags:
                description: List of tags.
                type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: string
        description:
          description: The description of the stream.
          type: string
        runId:
          description: The ID associated with the run modifying the stream.
          type: string
      required:
        - type
        - physicalName
        - sourceName
        - fields
        - schemaLocation
      example:
        type: STREAM
        physicalName: my-stream
        sourceName: my-stream-source
        schemaLocation: 'http://registry.example.com/schemas/ids/1'
        fields: []
        description: "My first stream dataset!"
...
    Dataset:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/DatasetId'
        type:
          description: The type of the dataset.
          type: string
        name:
          description: The **logical** name of the dataset.
          type: string
        physicalName:
          description: The **physical** name of the dataset.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was updated.
          type: string
          format: date-time
        namespace:
          type: string
          description: The namespace of the dataset.
        sourceName:
          description: The name of the source associated with the dataset.
          type: string
        fields:
          description: The fields of the dataset.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The data type of the field.
                type: string
              tags:
                description: List of tags.
                type: string
              description:
                description: The description of the field.
                type: string
        tags:
          description: List of tags.
          type: string
        lastModifiedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was last modified by a successful run.
          type: string
          format: date-time
        description:
          description:  The description of the dataset.
          type: string
        facets:
          $ref: '#/components/schemas/DatasetFacets'
        currentVersion:
          description: The current version of the dataset.
          type: string
          format: uuid
      example:
        id:
          namespace: my-namespace
          name: my-dataset
        type: DB_TABLE
        name: my-dataset
        physicalName: public.mytable
        createdAt: 2019-05-09T19:49:24.201361Z
        upodatedAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        sourceName: my-source
        fields: [
          {name': a, type: INTEGER, tags: []},
          {name': b, type: TIMESTAMP, tags: []},
          {name': c, type: INTEGER, tags: []},
          {name': d, type: INTEGER, tags: []}
        ]
        tags: []
        lastModifiedAt: null
        description: "My first dataset!"
        facets: {}
        currentVersion: "b1d626a2-6d3a-475e-9ecf-943176d4a8c6"
    DatasetList:
      type: object
      properties:
        datasets:
          items:
            $ref: '#/components/schemas/Dataset'
        totalCount:
          type: number
          description: The total number of datasets for the given namespace
      required:
        - datasets
...
    DatasetVersionList:
      type: object
      properties:
        versions:
          items:
            $ref: '#/components/schemas/DatasetVersion'
...
    CreatedJob:
      type: object
      properties:
        id:
          description: The ID of the job.
          type: object
          properties:
            namespace:
              type: string
              description: The namespace of the job.
            name:
              type: string
              description: The name of the job.
        type:
          $ref: '#/components/schemas/JobType'
        inputs:
          description: The set of **input** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        outputs:
          description: The set of **output** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        location:
          description: The URL of the job source code or artifact.
          type: string
          format: URL
        context:
          description: A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
          type: object
          deprecated: true
        description:
          description: The description of the job.
          type: string
        runId:
          description: An _optional_ run ID used to associate a job version to an existing job run.
          type: string
      required:
        - type
        - inputs
        - outputs
      example:
        type: BATCH
        inputs: [{ namespace: my-namespace, name: my-dataset}]
        outputs: []
        location: https://github.com/my-jobs/blob/124f6089ad4c5fcbb1d7b33cbb5d3a9521c5d32c
        context: {sql: 'SELECT * FROM mytable;'}
        description: My first job!
...
    JobVersionList:
      type: object
      properties:
        versions:
          items:
            $ref: '#/components/schemas/JobVersion'
...
    JobList:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        totalCount:
          type: number
          description: The total number of jobs for the given namespace
...